DRILL-6436: Storage Plugin to have name and context moved to AbstractStoragePlugin
