+++ contrib/storage-hive/core/src/main/java/org/apache/drill/exec/store/hive/HiveStoragePlugin.java
@@ -55,14 +55,12 @@ public class HiveStoragePlugin extends AbstractStoragePlugin {
 
   private final HiveStoragePluginConfig config;
   private HiveSchemaFactory schemaFactory;
-  private final DrillbitContext context;
-  private final String name;
   private final HiveConf hiveConf;
 
-  public HiveStoragePlugin(HiveStoragePluginConfig config, DrillbitContext context, String name) throws ExecutionSetupException {
+  public HiveStoragePlugin(HiveStoragePluginConfig config, DrillbitContext context, String name)
+      throws ExecutionSetupException {
+    super(context, name);
     this.config = config;
-    this.context = context;
-    this.name = name;
     this.hiveConf = createHiveConf(config.getHiveConfigOverride());
     this.schemaFactory = new HiveSchemaFactory(this, name, hiveConf);
   }
@@ -73,14 +71,6 @@ public class HiveStoragePlugin extends AbstractStoragePlugin {
 
   public HiveStoragePluginConfig getConfig() {
     return config;
-  }
-
-  public String getName(){
-    return name;
-  }
-
-  public DrillbitContext getContext() {
-    return context;
   }
 
   @Override
@@ -150,7 +140,7 @@ public class HiveStoragePlugin extends AbstractStoragePlugin {
       logger.warn("Schema factory forced close failed, error ignored", t);
     }
     try {
-      schemaFactory = new HiveSchemaFactory(this, name, hiveConf);
+      schemaFactory = new HiveSchemaFactory(this, getName(), hiveConf);
     } catch (ExecutionSetupException e) {
       throw new DrillRuntimeException(e);
     }
