+++ contrib/storage-jdbc/src/main/java/org/apache/drill/exec/store/jdbc/JdbcStoragePlugin.java
@@ -77,17 +77,14 @@ public class JdbcStoragePlugin extends AbstractStoragePlugin {
 
 
   private final JdbcStorageConfig config;
-  private final DrillbitContext context;
   private final DataSource source;
-  private final String name;
   private final SqlDialect dialect;
   private final DrillJdbcConvention convention;
 
 
   public JdbcStoragePlugin(JdbcStorageConfig config, DrillbitContext context, String name) {
-    this.context = context;
+    super(context, name);
     this.config = config;
-    this.name = name;
     BasicDataSource source = new BasicDataSource();
     source.setDriverClassName(config.getDriver());
     source.setUrl(config.getUrl());
@@ -429,8 +426,8 @@ public class JdbcStoragePlugin extends AbstractStoragePlugin {
 
   @Override
   public void registerSchemas(SchemaConfig config, SchemaPlus parent) {
-    JdbcCatalogSchema schema = new JdbcCatalogSchema(name);
-    SchemaPlus holder = parent.add(name, schema);
+    JdbcCatalogSchema schema = new JdbcCatalogSchema(getName());
+    SchemaPlus holder = parent.add(getName(), schema);
     schema.setHolder(holder);
   }
 
@@ -438,14 +435,6 @@ public class JdbcStoragePlugin extends AbstractStoragePlugin {
   @Override
   public JdbcStorageConfig getConfig() {
     return config;
-  }
-
-  public DrillbitContext getContext() {
-    return this.context;
-  }
-
-  public String getName() {
-    return this.name;
   }
 
   @Override
