+++ exec/java-exec/src/main/java/org/apache/drill/exec/store/dfs/FileSystemPlugin.java
@@ -62,7 +62,9 @@ public class FileSystemPlugin extends AbstractStoragePlugin {
   private final Configuration fsConf;
   private final LogicalPlanPersistence lpPersistance;
 
-  public FileSystemPlugin(FileSystemConfig config, DrillbitContext context, String name) throws ExecutionSetupException{
+  public FileSystemPlugin(FileSystemConfig config, DrillbitContext context, String name)
+      throws ExecutionSetupException{
+    super(context, name);
     this.config = config;
     this.lpPersistance = context.getLpPersistence();
